LIS-15 - Move atol error type source model to the separate folder
